moved all auth controllers to auth.contorller.js
