+++ routes/auth.route.js
@@ -1,10 +1,14 @@
 const express = require('express');
-const userCtrl = require('../controllers/user.controller');
+const authCtrl = require('../controllers/auth.controller');
 
 const router = express.Router(); // eslint-disable-line new-cap
 
-router.route('/forgot').post(userCtrl.forgotPassword);
+router.route('/forgot').post(authCtrl.forgotPassword);
 
-router.route('/reset/:token').post(userCtrl.resetPassword);
+router.route('/reset').get(authCtrl.reset);
+
+router.route('/reset').post(authCtrl.reset);
+
+router.route('/reset/:token').post(authCtrl.resetPassword);
 
 module.exports = router;
